// sourcecode/BTrak.Packages/snova-payroll-widgets/module/module-components/src/lib/payrollwidgetsandreports/models/payrollcalculationconfigurationsmodel.ts
export class PayRollCalculationConfigurationsModel{
    payRollCalculationConfigurationsId: string;
    periodTypeId: string;
    branchId: string;
    taxCalculationTypeName: string;
    payRollCalculationTypeId: string;
    payRollCalculationTypeName: string;
    branchName: string;
    isArchived: boolean;
    timeStamp: any;
    activeFrom: Date;
    activeTo: Date;
}