// sourcecode/BTrak.Packages/snova-payroll-widgets/module/module-components/src/lib/payrollwidgetsandreports/models/PayRollTemplateModel.ts
export class PayRollTemplateModel{
    payRollTemplateId: string;
    payRollName: string;
    payRollShortName: string;
    isRepeatInfinitly: boolean;
    islastWorkingDay: boolean;
    isArchived: boolean;
    timeStamp: any;
    frequencyId: string;
    infinitlyRunDate: Date;
    currencyId: string;
}